Added the GitClone class to bloom.git
The GitClone class/contextmanager provides a 
temporary clone of a git repository and will
only push the changes to that temporary clone if
the commands inside the with statement are
successful and the commit method is called.
This allows for the appearance of rolling back
changes to the repository on errors that occur
mid changes.