Bootstrap ROS 2 support for Bloom (#1)
* Continue using exceptions and conditionals for python 2/3
compatibility.

These changes will require review and testing before being accepted into Bloom.
They were hack-as-I-go changes required to run bloom inside a python3
virtualenv and I've done nothing so far to make sure they run fine under
python2.

* Specify correct package in error message.

This looks like a copy-paste mistake that could be cherry picked into upstream
bloom without the rest of these changes.

* Use different buildsystem arguments for debhelper based on package build_type.

This uses some inlined code from a hot'n'fresh PR to catkin_pkg and a
dictionary lookup to determine build options to send to debhelper.  At the
moment I am intentionally not adding continued support for catkin packages and
packages with no specified build type so that I can make sure to test that this
change in fact works for those packages.

* Move build options onto dh line.

A few debhelper builds failed when I tried to pass buildsystem arguments via the DH_OPTIONS environment variable.
A warning was shown for unknown options, and the value for an option was interpreted incorrectly. The debhelper manual page includes this:

		When using dh(1), it can be passed options that will be passed on
		to each debhelper command, which is generally better than using
		DH_OPTIONS.

and putting the options on the dh invocation line directly both fix the warning message and behave as expected.

These warnings are present on both the ROS 2 testfarm and the ROS buildfarm.