Resolve Python 2/3 encoding in git-bloom-import-upstream.
When running Bloom in Python 3, patched files are treated as binary but
the intermediate storage is still a python string. In order to be
written out as binary data it must be encoded as utf-8 bytes.

The stacktrace below is an example of the error on master.

    ==> git-bloom-import-upstream /tmp/tmppnro4mtb/upstream-master.tar.gz r2b2 --release-version 1.5.0 --replace
    Creating upstream branch.                                                                                                      │
    Importing archive into upstream branch...                                                                                      │
    Overlaying files from patched folder 'r2b2' on the 'master' branch into the 'upstream' branch...                               │
      Templating 'package.xml' into upstream branch...                                                                             │
    Traceback (most recent call last):                                                                                             │
      File "/home/steven/osrf/bloom2env/py3-ros1-upstream/bin/git-bloom-import-upstream", line 9, in <module>                      │
        load_entry_point('bloom', 'console_scripts', 'git-bloom-import-upstream')()                                                │
      File "/home/steven/osrf/bloom2env/bloom-upstream/bloom/commands/git/import_upstream.py", line 391, in main                   │
        args.replace)                                                                                                              │
      File "/home/steven/osrf/bloom2env/bloom-upstream/bloom/commands/git/import_upstream.py", line 319, in import_upstream        │
        import_patches(patches_path, patches_path_dict, 'upstream', version)                                                       │
      File "/home/steven/osrf/bloom2env/bloom-upstream/bloom/commands/git/import_upstream.py", line 217, in import_patches         │
        handle_tree(patches_path_dict, '', patches_path, version)                                                                  │
      File "/home/steven/osrf/bloom2env/bloom-upstream/bloom/commands/git/import_upstream.py", line 208, in handle_tree            │
        f.write(file_data)                                                                                                         │
    TypeError: a bytes-like object is required, not 'str'

There are likely other solutions to this issue, this was the one I used
for ROS 2's beta release but I can switch it up in review if there are
other suggestions.

The output from execute_command is not decoded as unicode if it comes
back as a `str`. If the value is a `str` containing unicode it will fail
to encode properly when writing the contents back out. This commit
avoids making broader changes the behavior in
`bloom.util.execute_command` which may wind up being the ultimate
resolution in order to avoid breaking other parts of bloom which may
rely on the current behavior even though it's brittle around utf-8
strings.

This was tested using eProsima/Fast-RTPS from ROS 2 beta2 (r2b2) because
it's the only package I know directly of with utf-8 encoded author and
maintainer names. I'd love to test against a ROS kinetic or lunar
package as well in order to have a pure ROS 1 test environment.

The python versions tested were
- 2.7.12
- 3.5.2