Don't override `install-scripts` from setup.cfg. (#2)
* Support converting dicts through unicode.

* Remove empty exportvars.

There's a default setting when none are specified it's better just to
use that.

* Use a dict for exportvars instead of a list of tuples.

This makes it easier to manipulate exported variables after the fact,
which we're going to do to handle conditional `--install-scripts`.

This was what I tried first but since the unicode conversion routine
didn't support dictionaries I ended up going the tuple route. Now that I
actually needed to access specific variables as a dictionary it made
more sense to add dict support and use it than to continue to work
around it.

* Check ament_python packages for install-scripts in setup.cfg.

This looks for a `setup.cfg` file in the package path, tries to parse it
with configparser, and if the `install.install-scripts` is present, will
not override it in the debian/rules. If there is no setup.cfg or there
is no install.install-scripts option it will default to installing
scripts into the ROS prefix `bin` directory. For example
`/opt/ros/r2b2/bin`

* Remove quotes so they are not part of the literal argument.

So, after a conversation with mikaelarguedas about these quotes where I
described that while they aren't necessary in a Makefile they would be
in POSIX shell and thus having them is Just Good Sense :tm: it of course turns out that Make preserves the quotes as part of the
literal value and thus they gum up the usage when the variable is
exported to the shell.

Removing the quotes resolves the problem.