Consolidate ROS 2-related build_type changes.
Some of the build-type specific changes were made in a function
initially called `set_debhelper_options` and some were made inline in
the substitution generation function. This brings all of those changes
together and renames the function `set_build_type_options`.

If we wanted to, we could merge this directly into upstream bloom (once
support for the catkin build type is added) and call it done.

But I think this is pretty messy and I'm at the point where I think that
I want separate templates for each build type. That work is ongoing in
the separate-build-types branch which probably needs a messy rebase in
order to be accurately compared to this one.