Print 2FA message when encountering GitHubAuthException.
The existing error handling code has been refactored to print the
multi-factor authentication prompt either when the returned
GitHubException has an attached response with the 401 status code or
when a GitHubAuthException is raised.

It's possible that the old behavior is no longer needed and can be
removed rather than left in place but I'm not familiar enough with the
urllib2 API to say.