Fix python 3 support for pull request generation. (#489)
Calling `.encode('utf-8')` on a python2 unicode object encodes it into a
utf-8 str but calling it on a python3 str encodes that str into a bytes
object.

We want python str objects in both cases to be serialized into the
distribution yaml files. To achieve that we filter the modified
distribution repo dict recursively for unicode keys and values
and encoding them to utf-8 strings.

This also fixes some issues in the GitHub library related to Python 3.