Use bloom distribution info for conditional context.
Rather than using the local environment, this commit updates the
conditional evaluation to use a context created from distribution data
already available to bloom.

It's probably reasonable to want to override this somehow, either with
explicit context in the tracks.yaml, local environment variables, or
both. But I think this is the least-surprising thing to do, the only
critical environment variable missing still is the ROS_PYTHON_VERSION
which will need a sensible source.