Disable automagic byte compilation in RPM spec files
The BRP Python bytecompiler will always use the sytem's default Python
interpreter, which may not be the interpreter we're targeting. Safest
option is to disable the automagic byte compilation altogether.

Note that this doesn't mean that python files won't ever be compiled,
it just means that the catch-all policy at the end of the process won't
attempt to compile anything which hasn't already been compiled.

Signed-off-by: Scott K Logan <logans@cottsay.net>