Use 'cmake3' macro in RPM spec files
This change shouldn't modify the behavior in Fedora, where all current
releases define 'cmake3' to be the same as 'cmake'. In RHEL 7, where
cmake 2 is the default, we need to use the 'cmake3' macro to use the
supplamental 'cmake3' executable instead of the system default.

All current ROS releases except Indigo have a *minimum* cmake
requirement of 3.

Signed-off-by: Scott K Logan <logans@cottsay.net>