get_package_data return list, reuse it
Notice that the `get_package_data` method will return a list.
so it's clear that the logic `detect_branches` method in ReleaseGenerator and RosReleaseGenerator is same.
In this commit, I reuse this function and put it into ReleaseGenerator's `handle_arguments`
method so that we can safely delete the duplicate part in RosReleaseGenerator.