[VCPKG] add ros vcpkg generator
this generator follow the style of old ros generators

But notice that I call VcpkgGenerator.get_subs_hook in the end of
RosVcpkgGenerator's get_subs_hook, this way follow the Bloom style.
If we don't do this way, the appended BuildDepends and Depends won't be
updated to VcpkgGenerator.get_subs_hook, and will cause depend losed problem.