Add ROS_PYTHON_VERSION to dependency conditionals. (#551)
* Raise rosdistro requirement to include new index-v4 changes.

The python_version key is required to supply ROS_PYTHON_VERSION to
conditional dependencies.

* Add API method for retrieving the distribution's python version.

* Add ROS_PYTHON_VERSION to conditional context.

* Fix pep8 conformance.

* Explicitly note when the python_version key is missing.

This feature will cause interruptions for bloom users whose version 4
rosdistro indexes lack this field. But I think this is preferred to a
silent or even vocal acceptance of a possibly invalid release state due
to missing conditional variables.