Provide -devel, -doc and -runtime virtual packages in RPM
There is nothing that will currently take advantage of these virtual
packages, but providing them now would allow us to implement a mechanism
to take advantage of them in the future without re-blooming everything.

The virtual packages which an RPM package provides are not generally
exposed in any package lists, so this shouldn't cause any confusion.

Since the current behavior is to include all of a ROS package's files
(including runtime, development and documentation) in a single package,
these virtual packages are not incorrect.