Provide group membership information in RPM packages (#564)
This is not a complete implementation of group support in RPMS. The
package.xml format does not contain enough information for us to specify
the specifics of a group dependency. However, we do have enough
information to express membership.

Nothing will automatically take advantage of this new dependency
information right now, but it will open the door to a mechanism in the
future without re-blooming all group members.

We could also take advantage of this information in the manual patching
we're currently doing to express dependencies on groups.