Add non-interactive option to git-bloom-release command.
This adds a noninteractive option to git-bloom-release. Most of the time
this command is often non-interactive but several fields in the bloom
config can be set to :{ask} which would normally block with a prompt.

With this patch an error message will be raised instead if the :{ask}
option is used.