Fix conditional evaluation of Conflicts/Replaces
It seems that the conditional logic was only getting applied to the
conflicts and replaces keys to prevent resolution, but the condition
isn't checked when adding the relationship to the package, resulting in
a KeyError when a resolved name for the target package isn't available.

This change treats the conflicts and replaces lists the same way as the
run and build dependencies.