Add CMAKE_PREFIX_PATH to ament_cmake and cmake templates
This CMake variable is already set for the `catkin` template, but is not
currently set for the `ament_cmake` and plain `cmake` templates.

The `CMAKE_INSTALL_PREFIX` variable seems to be sufficient to allow
CMake to find packages which are installed under that path, however when
a package is present in a system location as well as under the target
installation prefix, preference is given to the system package (given
that it satisfies any version requirements specified).

Setting `CMAKE_PREFIX_PATH` should give preference to the packages under
that path when a package is also available as a system package and both
satisfy any version requirements.