Create and use a 'fake' rosdep cache for tests (#614)
The idea here is to decouple Bloom from the live rosdistro index and
rosdep database. This will also isolate bloom from the rosdep state that
the system has, and makes the tests function without any network access.
Since the system rosdep state isn't a dependency any more, the tests can
be run even if 'rosdep init' has not been invoked on the system, meaning
that the tests can now be executed without sudo rights.

This change also significantly reduces the time it takes to run the
tests. Part of that is due to the 'BLOOM_SKIP_ROSDEP_UPDATE', which
suppresses Bloom's automatic rosdep update invocation. This isn't
necessary for the 'fake' rosdep cache to work, but it makes things a bit
quicker.