Add a template substitution for the ROS distro (#612)
The most common patching error I've seen in Bloom repositories manifests
from carrying forward a patch from a previous track to a new one, when
that patch has a hard-coded ROS distribution name in it. Unfortunately,
this error often goes unnoticed when the patch modifies the weak
dependencies of a package.

This substitution will allow us to roll the patches forward unmodified.