Ignore --skip-keys arguments when inspecting actions
The prompt to update a package's action list is useful for encouraging
maintainers to take on new generators, but it seems that many
maintainers are accepting an update to the default action list without
trying to figure out why the list differs to begin with.

The --skip-keys argument was added to the RPM generator to work around
some non-critical missing dependencies in various packages, but the
argument keeps getting removed. This change doesn't prevent the argument
from being removed when a legitimate difference is detected, but it does
prevent the prompt from being raised when the action list differs only
by --skip-keys arguments.