Use basic authentication with new personal access tokens.
New personal access tokens on GitHub are not created as oauth tokens and
are not usable with token authentication. Instead they're used as the
password for basic authentication.

This is a minimal change to ensure that bloom can continue to work with
these new tokens.

I'd like to spend some time soon reviewing and refreshing the GitHub API
behavior in order to improve the performance and reliability of the
automatic pull request workflow and now that multi-factor authentication
is widely deployed for GitHub, improve the messaging around first time
setup with Bloom.