Collect manually created token rather than attempting to create one.
Since multi-factor authentication has become paramount for GitHub and
similar sites I hope that most of Bloom's users encounter the failure to
automatically create a token on their behalf.

This updates the authentication failure mode to provide instructions for
creating a token via the GitHub web UI and copy it into Bloom's
configuration.

The authorizations API previously used by bloom is not documented
currently so I didn't have a reference for trying to continue using it
with MFA.