Add a happy-fast check for a guessed fork name.
The ros/rosdistro repository has accumulated many forks and the current
bloom strategy of listing them all then finding the one belonging to the
current GitHub user takes time.

Most of the time a forked repository has the same name as the parent.
This PR checks that the target repository and a user repository with the
same name (if it exists) share the same "source" and are thus in the
same fork network so pull requests can be created between them.

If the fork exists with the right name this check is now almost
instantaneous on my local internet connection so this could save a lot
of seconds when working on releases to ros/rosdistro and since it
creates only two additional API requests it should not slow down the
exhaustive search by very much.