Fix token authentication error.
The new personal access tokens from GitHub for which support was first
added in bloom in #627 use basic authentication like a username and
password rather than the OAuth token authentication used by OAuth tokens
and previous iterations of the personal access token. When adding
support for this it seems like I missed a trick and as a result the
authentication header was being sent as
`Basic b'base64encodeduserandtoken'` rather than
`Basic base64encodeduserandtoken`.

I have to do an experimental bisect to determine why this issue is only
showing up now. Perhaps it's been an issue since the jump to python3 but
since basic authentication was only ever used as a springboard for token
auth it was overlooked when reported.