Shallow clone rosdistro for creating PRs
The history in rosdistro is _really_ long, and it can take quite a while
to pull it when creating PRs. Since we're just going to make a single
commit on top, there's really no reason to pull the full history.

This makes creating the PR faster - even on fast internet connections.