Begin RPM build directories with '.'
The standard practice for CMake build directories in RPM packages is to
create a unique subdirectory within the source tree. This clashes with
many of the linters in ROS and ROS 2, which recursively process the
package from the root of the package directory, leading to numerous
linter failure messages during the RPM test phase.

It appears that the majority of these linters will ignore a 'hidden'
directory which begins with a '.', so naming our build directory as such
may help prevent the linters from crawling where we don't want them to.