Conditionally exclude test dependencies
In the debian generator, we can decorate the BuildDepends to omit them
if the `nocheck` flag is set.

For RPMs, there isn't a standard mechanism for skipping tests but we
actually created a build flag for skipping tests, so the only work here
was to wire that custom flag into the dependency enumeration.