Refactor CI platforms.
The setup-python action has dropped 2.7 on macOS.
Aside from that, while it's possible for the exclude clause to be more
succinct in practice it has been more difficult to reason about.

The changes match those of [catkin_pkg#318].
Python 3.7 testing on ubuntu-latest is meant to cover Debian Buster.

Beyond that the full changes are enumerated below

CI platforms before:
* ubuntu-18.04
  * python 2.7
  * python 3.6
  * python 3.7
  * python 3.8
  * python 3.9
* ubuntu-20.04
  * python 3.7
  * python 3.8
  * python 3.9
* macos-latest
  * python 2.7
  * python 3.7
  * python 3.8
  * python 3.9

CI platforms after:
* ubuntu-18.04
  * python 2.7
  * python 3.6
* ubuntu-20.04
  * python 3.7
  * python 3.8
  * python 3.9
* macos-latest
  * python 3.7
  * python 3.8
  * python 3.9

[catkin_pkg#318]: https://github.com/ros-infrastructure/catkin_pkg/pull/318