Export BLOOM_NON_INTERACTIVE to track actions environment.
If git-bloom-release is run non-interactively then that should be
propagated to the actions command so that actions are not blocked on
interactivity.

This patch also makes the minor optimization of only making one copy of
the environment to pass to the actions run. I don't think this makes a
major difference in performance but it is one less thing to worry about
and environment changes from children should not persist to subsequent
children or "bubble up" to the parent process.