Exit early on patch failure if bloom is being run non-interactively. (#668)
* Export BLOOM_NON_INTERACTIVE to track actions environment.

If git-bloom-release is run non-interactively then that should be
propagated to the actions command so that actions are not blocked on
interactivity.

This patch also makes the minor optimization of only making one copy of
the environment to pass to the actions run. I don't think this makes a
major difference in performance but it is one less thing to worry about
and environment changes from children should not persist to subsequent
children or "bubble up" to the parent process.

* Exit early on patch failure if bloom is being run non-interactively.

* Indentation.

* Use isatty() instead of requiring an environment variable.

* Revert "Export BLOOM_NON_INTERACTIVE to track actions environment."

This reverts commit a9fe30c29e9c1b5748d4379dc48875cb07b91b0a.
The updated implementation relies on using isatty for interactive
terminal detection rather than an exported environment variable.