Update deb platforms and drop python2 packaging.
Following in the wake of catkin_pkg, we're dropping python2 support now
that it's no longer the default on any active distribution.

Dropped:
* Ubuntu Bionic (18.04 LTS)
* Debian Buster (oldoldstable)
* Debian Bullseye (oldstable)

Added:
* Ubuntu Noble (24.04 LTS pre-release)
* Debian Trixie (testing)

Retained:
* Ubuntu Focal (20.04 LTS)
* Ubuntu Jammy (22.04 LTS)

I have also raised the minimum python3 version to 3.6 as we are only
testing back to 3.6 and RHEL 8 has the oldest python distribution (3.6)
of any supported platform.