Fix no remote check.
This check has likely never worked as intended. I'm not aware of any
case where `len()` returns a negative nubmer. The described error seems
like the intent was to check for an empty list of remotes not a somehow
negative number of remotes.

When updating this check, there are several test failures which  i have
not triaged. It's possible that now-expected behavior is incompatible
with this remote check or that our test fixtures need to be updated so
they don't cause this error.